+++ src/main/java/org/logstash/snmp/SnmpClient.java
@@ -306,8 +306,9 @@ public class SnmpClient implements Closeable {
             trapEvent.put("error_status_text", pdu.getErrorStatusText());
             trapEvent.put("error_index", pdu.getErrorIndex());
             if (pdu instanceof ScopedPDU) {
-                trapEvent.put("context_engine_id", ((ScopedPDU)pdu).getContextEngineID());
-                trapEvent.put("context_name", ((ScopedPDU)pdu).getContextName());
+                final ScopedPDU scopedPDU = (ScopedPDU) pdu;
+                trapEvent.put("context_engine_id", String.valueOf(scopedPDU.getContextEngineID()));
+                trapEvent.put("context_name", scopedPDU.getContextName());
             }
         } else if (pdu instanceof PDUv1) {
             final PDUv1 pdUv1 = (PDUv1) pdu;